requirements.txt
Django==3.0.5
djangorestframework==3.10.3
psycopg2-binary==2.8.4
